requirements.txt
streamlit==1.42.0
pymupdf==1.25.1
arxiv==2.1.3
sentence-transformers==3.4.1
chromadb==0.6.3
langchain==0.3.20
langchain-community==0.3.19
langchain-groq==0.2.0
langgraph==0.3.5
python-dotenv==1.0.1
hf_xet